requirements.txt
click==7.1.2
colorama==0.4.4
pillow==8.1.2
requests==2.25.1
rich==9.13.0